+++ packages/terra-application-header-layout/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.3.0 - (March 20, 2019)
+------------------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
 3.2.0 - (February 26, 2019)
 ------------------
 ### Added