+++ packages/terra-application-layout/CHANGELOG.md
@@ -3,7 +3,11 @@ Changelog
 
 Unreleased
 ----------
+
+4.6.0 - (March 20, 2019)
+------------------
 ### Changed
+* Update Jest tests to use terra-enzyme-intl
 * Changed demo mock colors to be themable
 
 ### Added