+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.10.0 - (March 20, 2019)
+------------------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
 3.9.0 - (March 12, 2019)
 ------------------
 ### Changed