+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+2.6.0 - (March 20, 2019)
+------------------
 ### Changed
 * Doc site 'modal on modal' example heading level changed to correct Section 508 error